Introduction of specialized operators for combining multidimensional tables.
New specialized operators should be written in files
multidim/operators4MultiDim.h and multidim/operators4MultiDim.tcc. For the
moment, only the additions, subtractions, multiplications and divisions
for multiDimArrays have been inserted (the real codes of these operators
are written in subdirectory multidim/patterns). These functions can be
called directly by the user. In addition, operators +, -, *, / have been
added to file multidim/operators4MultiDim.h. These select, among all the
functions defined for combining multidimensional tables, that which is the
most appropriate (specialized) for the operation to perform, hence ensuring
fast computations. To know about the appropriate functions, an
OperatorRegister has been added as well and this one is created by
multiDimDecorators. Hence, the first time a multiDimDecorator or
one of its descendants is created, all the operators on multidims will be
registered. If no decorator is created, function operators4MultiDimInit<T>() 
should be used to register all the operators that should be known to
operators +, -, /, *. Finally, names have been added to all the
multidimensional table types. These are used to select the appropriate
functions to call within operators +, -, /, *.



git-svn-id: https://alm.lip6.fr/svn/agrum/trunk@302 fa2884e2-1fde-468d-ae8a-4cc2b5279982