first draft for the schedule class.
Basically, a Schedule is a set of operations to perform with
some constraints on the order in which the operations should
be performed (e.g., when computing (f1 + f2) + f3, we must
compute f1 + f2 before adding the result with f3).

The purpose of the Schedule class is not to perform the execution
of the operations because different schemes could be used. Its purpose
is just to set the necessary partial execution order between the 
operations.



git-svn-id: https://alm.lip6.fr/svn/agrum/trunk@524 fa2884e2-1fde-468d-ae8a-4cc2b5279982