adding a multiDimCombineAndProject that efficiently uses MultiDimCombination and MultiDimProjection to perform operations on junction tree cliques.
Adding methods nbOperations and memoryUsage to MultiDimCombination, 
multiDimProjection and multiDimCombineandProject: these methods allow to
assess the time and space complexity of inference without actually performing
the latter.

adding methods nbOperations and memoryUsage to the Schedule classes. Now, there
is a SchedulerBasic that can actually perform the computations of a given schedule.



git-svn-id: https://alm.lip6.fr/svn/agrum/trunk@533 fa2884e2-1fde-468d-ae8a-4cc2b5279982